Correct contract mistake
Made a mistake in the code, but didn't find it until I moved to the
main library. Since the version number is in lockstep with the main
library, I'm introducing a weird convention of an extra minor version
when the contract needs to be updated but not in a way that changes
the main library's behavior.